requirements.txt
aiohttp==1.3.3
APScheduler==3.6.3
asgiref==3.3.0
assertpy==1.0
astroid==2.3.3
async-timeout==3.0.1
blinker==1.4
Brotli==1.0.9
certifi==2019.9.11
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
cycler==0.10.0
dataclasses==0.7
defusedxml==0.6.0
demjson==2.2.4
diff-match-patch==20181111
Django==2.2.17
django-apscheduler==0.3.0
django-cors-headers==3.2.0
django-crispy-forms==1.8.1
django-debug-toolbar==3.1.1
django-filter==2.1.0
django-formtools==2.1
django-guardian==2.1.0
django-import-export==1.2.0
django-redis==4.10.0
django-rest-swagger==2.2.0
django-reversion==3.0.4
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
drf-yasg==1.17.1
et-xmlfile==1.0.1
Flask==1.1.2
future==0.18.2
h11==0.11.0
h2==3.2.0
hpack==3.0.0
httplib2==0.14.0
hyperframe==5.2.0
idna==2.8
inflection==0.4.0
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
jsonpath==0.82
kaitaistruct==0.8
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
ldap3==2.7
Markdown==3.0.1
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mitmproxy==5.2
multidict==4.7.5
mysqlclient==1.4.6
numpy==1.18.2
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==3.0.2
packaging==20.3
passlib==1.7.4
Pillow==6.2.1
protobuf==3.11.3
psycopg2==2.8.4
publicsuffix2==2.20191221
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.9.4
pydivert==2.1.0
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyperclip==1.8.0
python-dateutil==2.8.1
pytz==2018.9
PyYAML==5.1.2
raven==6.10.0
redis==3.3.11
requests==2.22.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.13.0
sortedcontainers==2.1.0
sqlparse==0.4.1
tablib==0.14.0
tornado==6.0.4
typed-ast==1.4.0
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.25.7
urwid==2.0.1
Werkzeug==1.0.1
win32-setctime==1.0.1
wrapt==1.11.2
wsproto==0.15.0
xlrd==1.2.0
XlsxWriter==1.2.6
xlwt==1.3.0
yarl==0.9.8
zstandard==0.13.0
